When answering a call, stop ringing even if network takes too long to connect.
When answering a call, we tell the Ringer to stop ringing. However,
CallNotifier receives a continuous stream of PHONE_INCOMING_CALL events
which restarts the ringer.  The result is that the phone will start
ringing again if RIL takes too long to actually connect the call. Instead
of stopping the ringer once, this CL changes the code to "silence" the
ringer indefinitely for the current call. This piggy backs the same code
which silences the ringer when the user opts to silence the call instead
of answering it.

bug: 7439536
Change-Id: Ie2a1c83e147435821fca4d314414075c305d6aa3